% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag-skeleton.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Extract adjacency matrix from a DAG skeleton.}
\usage{
adjacency_matrix(.dag_skeleton)
}
\arguments{
\item{.dag_skeleton}{The PC DAG skeleton object.}
}
\value{
Outputs an adjacency matrix of the DAG skeleton.
}
\description{
Is generally a wrapper around calls to \code{\link[igraph:get.adjacency]{igraph::get.adjacency()}} and
\code{\link[igraph:igraph.from.graphNEL]{igraph::igraph.from.graphNEL()}}. Transforms from a GraphNEL object in igraph.
}
\examples{

\dontrun{
library(dplyr)
skeleton_estimate <- simulated_data \%>\%
select(contains("metabolite")) \%>\%
NetCoupler:::pc_skeleton_estimates()

NetCoupler:::adjacency_matrix(skeleton_estimate)
}

}
