% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{as_edge_tbl}
\alias{as_edge_tbl}
\title{Convert network graph to edge table.}
\usage{
as_edge_tbl(.network)
}
\arguments{
\item{.network}{Network graph from e.g. \code{\link[=nc_estimate_network]{nc_estimate_network()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble}, with at least two columns:
\itemize{
\item \code{source_node}: The starting node (variable).
\item \code{target_node}: The ending node (variable) that links with the source node.
}
}
\description{
\lifecycle{experimental}
}
\seealso{
See \link{nc_model_estimates} for examples on using NetCoupler.
}
