% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-links.R
\name{classify_options}
\alias{classify_options}
\title{Classification options for direct, ambigious, and no effect.}
\usage{
classify_options(
  single_metabolite_threshold = 0.05,
  network_threshold = 0.1,
  direct_effect_adjustment = NA
)
}
\arguments{
\item{single_metabolite_threshold, network_threshold, direct_effect_adjustment}{See \code{nc_estimate_links} for details.}
}
\value{
List with options for the classification.
}
\description{
Classification options for direct, ambigious, and no effect.
}
