% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{compute_adjacency_graph}
\alias{compute_adjacency_graph}
\title{Compute the adjacency matrix of the graph with the data.}
\usage{
compute_adjacency_graph(data, edge_tbl)
}
\arguments{
\item{data}{The data that was also used to generate the \code{\link[=nc_estimate_network]{nc_estimate_network()}}.}

\item{edge_tbl}{The network graph object of the metabolic variable network.}
}
\value{
Outputs an \code{igraph} object from \code{\link[igraph:graph_from_adjacency_matrix]{igraph::graph_from_adjacency_matrix()}}.
}
\description{
\lifecycle{experimental}
}
\keyword{internal}
