% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{compute_adjacency_matrix}
\alias{compute_adjacency_matrix}
\title{Extract adjacency matrix from a DAG skeleton.}
\usage{
compute_adjacency_matrix(dag_skeleton)
}
\arguments{
\item{dag_skeleton}{The PC DAG skeleton object.}
}
\value{
Outputs an adjacency matrix of the DAG skeleton.
}
\description{
\lifecycle{experimental}

Is generally a wrapper around calls to \code{\link[igraph:as_adjacency_matrix]{igraph::as_adjacency_matrix()}} and
\code{\link[igraph:graph_from_graphnel]{igraph::graph_from_graphnel()}}. Transforms from a GraphNEL object in igraph.
}
\keyword{internal}
