% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-netcoupler-functions.R
\name{getExp.coef.out}
\alias{getExp.coef.out}
\title{Extract coefficients of effect estimates for outcomes.}
\usage{
getExp.coef.out(object, metabolite, DE = NA)
}
\arguments{
\item{object}{Output of net.coupler}

\item{metabolite}{Specific metabolite to evaluate}

\item{DE}{Direct effect ...}
}
\value{
Outputs a data frame of effect estimates.
}
\description{
\lifecycle{soft-deprecated}
}
\details{
This function produces a table of effect estimates of all (some)
network-variables on an outcome (time-to-event) for all possibles causal
models based on conditional independence criteria encoded in the
input-network => MULTISET OF POSSIBLE EFFECTS PER EXPOSURE-OUTCOME PAIR;
network-variables of interest are selected by indicating the variable-names
as character vector.
}
