% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-netcoupler-functions.R
\name{getExp.coef.permetabolite}
\alias{getExp.coef.permetabolite}
\title{Extract the exposure coefficients per outcome from the netcoupler outcome estimation.}
\usage{
getExp.coef.permetabolite(object, metabolite, DE = NA)
}
\arguments{
\item{object}{Output of net.coupler.out}

\item{metabolite}{Specific metabolite to evaluate}

\item{DE}{Direct effect...}
}
\value{
Outputs a data frame of estimates.
}
\description{
\lifecycle{soft-deprecated}
}
