% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{nc_adjacency_graph}
\alias{nc_adjacency_graph}
\title{Title}
\usage{
nc_adjacency_graph(.data, .graph)
}
\arguments{
\item{.graph}{}
}
\value{

}
\description{
Title
}
\examples{

library(dplyr)
metabolite_data <- simulated_data \%>\%
  select(starts_with("metabolite"))
network <- simulated_data \%>\%
  select(contains("metabolite")) \%>\%
  nc_create_network()
nc_adjacency_graph(metabolite_data, network)
}
