% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{nc_adjacency_matrix}
\alias{nc_adjacency_matrix}
\title{Extract adjacency matrix from a DAG skeleton.}
\usage{
nc_adjacency_matrix(.dag_skeleton)
}
\arguments{
\item{.dag_skeleton}{The PC DAG skeleton object.}
}
\value{
Outputs an adjacency matrix of the DAG skeleton.
}
\description{
\lifecycle{experimental}

Is generally a wrapper around calls to \code{\link[igraph:as_adjacency_matrix]{igraph::get.adjacency()}} and
\code{\link[igraph:graph_from_graphnel]{igraph::igraph.from.graphNEL()}}. Transforms from a GraphNEL object in igraph.
}
\examples{

library(dplyr)
skeleton_estimate <- simulated_data \%>\%
  select(contains("metabolite")) \%>\%
  nc_create_network()

nc_adjacency_matrix(skeleton_estimate)

}
