% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{nc_create_network}
\alias{nc_create_network}
\title{Create an estimate of the metabolic network skeleton.}
\usage{
nc_create_network(.data, .alpha)
}
\arguments{
\item{.data}{Data of the metabolic variables.}

\item{.alpha}{The alpha level to set.}
}
\value{
Outputs a DAG skeleton.
}
\description{
Main NetCoupler network creator. Estimates the skeleton based on a family of
DAGs without specifying the direction of edges.
}
\examples{
library(dplyr)
simulated_data \%>\%
select(contains("metabolite")) \%>\%
nc_create_network()
}
