% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-estimates.R
\name{nc_exposure_estimates}
\alias{nc_exposure_estimates}
\title{Estimating pathways to an estimated DAG from an exposure.}
\usage{
nc_exposure_estimates(
  .data,
  .graph,
  .exposure,
  .adjustment_vars,
  .model_function,
  .exponentiate = FALSE,
  ...
)
}
\arguments{
\item{.data}{A data.frame that contains the data with the metabolic variables
and the outcome.}

\item{.graph}{The estimated graph skeleton obtained from \code{\link[=nc_create_network]{nc_create_network()}}.}

\item{.exposure}{Character. The exposure variable of interest.}

\item{.adjustment_vars}{Character vector. The variables to adjust for in the model.}

\item{.model_function}{A function object. The function to use for the modeling.}

\item{.exponentiate}{Logical. Default is false. Whether to exponentiate the
model results output, usually necessary for e.g. logistic regression
models.}

\item{...}{Options to pass to the model function.}
}
\value{
Outputs a \link[tibble:tibble-package]{tibble} with all the models computed and their
estimates.
}
\description{
\lifecycle{experimental}
}
\details{
This algorithm estimates direct effects of a predefined exposure on each
network-variable for all causal models that agree with the input-network:
models are adjusted for all possible combinations of direct neighbors
(==variables in the adjacency set) -> Output is a multiset of possible
effects
}
\examples{

library(dplyr)
library(survival)
metabolite_network <- simulated_data \%>\%
  select(matches("metabolite")) \%>\%
  nc_create_network()
simulated_data \%>\%
  nc_exposure_estimates(
    .graph = metabolite_network,
    .exposure = "exposure",
    .adjustment_vars = "age",
    .model_function = lm
   )

}
