% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-netcoupler-functions.R
\name{nc_make_network}
\alias{nc_make_network}
\title{Main NetCoupler network creator.}
\usage{
nc_make_network(.dataset, .alpha, .network_variables)
}
\arguments{
\item{.alpha}{The alpha level to set.}

\item{.network_variables}{Variables to set for the nextwork.}

\item{.data}{Input data.}

\item{.filter_variable}{Which variables to filter.}

\item{.filter_value}{The filter value.}

\item{.adjustment_variable}{Variables to adjust for.}
}
\value{
Outputs a DAG skeleton.
}
\description{
Main NetCoupler network creator.
}
