% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-estimates.R
\name{nc_outcome_estimates}
\alias{nc_outcome_estimates}
\title{Estimating pathways to an outcome from an estimated DAG.}
\usage{
nc_outcome_estimates(
  .data,
  .graph,
  .outcome,
  .adjustment_vars,
  .model_function,
  .exponentiate = FALSE,
  ...
)
}
\arguments{
\item{.data}{A data.frame that contains the data with the metabolic variables
and the outcome.}

\item{.graph}{The estimated graph skeleton obtained from \code{\link[=nc_create_network]{nc_create_network()}}.}

\item{.outcome}{Character. The outcome variable of interest.}

\item{.adjustment_vars}{Character vector. The variables to adjust for in the model.}

\item{.model_function}{A function object. The function to use for the modeling.}

\item{.exponentiate}{Logical. Default is false. Whether to exponentiate the
model results output, usually necessary for e.g. logistic regression
models.}

\item{...}{Options to pass to the model function.}
}
\value{
Outputs a \link[tibble:tibble-package]{tibble} with all the models computed and their
estimates.
}
\description{
\lifecycle{experimental}
}
\details{
This algorithm estimates direct effect of a predefined exposure
(network-variable) on time-to-event for all causal models that agree with the
input-network: Cox prop. hazards regression models are used to estimate the
efect of all network-variables on survival time adjusted for all possible
combinations of direct neighbors (adjacency set) -> Output is a multiset of
possible causal effects.
}
\examples{

library(dplyr)
library(survival)
metabolite_network <- simulated_data \%>\%
  select(matches("metabolite")) \%>\%
  nc_create_network()
simulated_data \%>\%
  nc_outcome_estimates(
    .graph = metabolite_network,
    .outcome = "survival::Surv(survival_time, case_status)",
    .adjustment_vars = "age",
    .model_function = survival::coxph,
    .exponentiate = TRUE
   )

}
