% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{nc_partial_corr_matrix}
\alias{nc_partial_corr_matrix}
\title{Estimate Pearson's partial correlation coefficients.}
\usage{
nc_partial_corr_matrix(.data)
}
\arguments{
\item{.data}{Input data of metabolic variables as matrix or data.frame.}
}
\value{
Outputs a matrix of partial correlation coefficients.
}
\description{
\lifecycle{experimental}

This function is a wrapper around \code{\link[ppcor:pcor]{ppcor::pcor()}} that extracts correlation
coefficient estimates, then adds the variable names to the column and row names.
}
\examples{

library(dplyr)
simulated_data \%>\%
  select(contains("metabolite")) \%>\%
  nc_partial_corr_matrix()

}
