% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{nc_plot_outcome_estimation}
\alias{nc_plot_outcome_estimation}
\alias{nc_plot_exposure_estimation}
\title{Plot of the outcome or exposure side model estimation.}
\usage{
nc_plot_outcome_estimation(
  .tbl,
  .graph,
  .tbl_model,
  .edge_label_threshold = 0.2
)

nc_plot_exposure_estimation(
  .tbl,
  .graph,
  .tbl_model,
  .edge_label_threshold = 0.2
)
}
\arguments{
\item{.tbl}{The original data, with the metabolic variables that have been
standardized.}

\item{.graph}{The graph object created from \code{nc_estimate_network()}.}

\item{.tbl_model}{The tibble of the model results obtained from
\code{nc_classify_effects()}.}

\item{.edge_label_threshold}{Threshold to pass for the value to be added to
the edge label.}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot2} object showing the model
estimation results linked with the network graph.
}
\description{
\lifecycle{experimental}

Plots the results of the effect classification based on the model estimation,
linking the results into the network graph.
}
\section{Functions}{
\itemize{
\item \code{nc_plot_exposure_estimation}: Plots the exposure side estimation.
}}

