% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCoupler-package.R
\docType{package}
\name{NetCoupler-package}
\alias{NetCoupler}
\alias{NetCoupler-package}
\title{NetCoupler: Inference of Causal Links Between a Network and an External Variable}
\description{
The NetCoupler algorithm identifies potential direct effects of correlated, high-dimensional variables formed as a network with an external variable. The external variable may act as the dependent/response variable or as an independent/predictor variable to the network.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NetCoupler/NetCoupler}
  \item Report bugs at \url{https://github.com/NetCoupler/NetCoupler/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

Authors:
\itemize{
  \item Clemens Wittenbecher \email{Clemens.Wittenbecher@dife.de} (\href{https://orcid.org/0000-0001-7792-877X}{ORCID})
}

Other contributors:
\itemize{
  \item Fabian Eichelmann \email{Fabian.Eichelmann@dife.de} [contributor]
  \item Helena Zacharias \email{helena.zacharias@helmholtz-muenchen.de} [contributor]
  \item Daniel Ibsen \email{dbi@ph.au.dk} (\href{https://orcid.org/0000-0002-7038-4770}{ORCID}) [contributor]
}

}
\keyword{internal}
