% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag-skeleton.R
\name{pc_skeleton_estimates}
\alias{pc_skeleton_estimates}
\title{Estimate order-independent PC-stable skeleton of a DAG.}
\usage{
pc_skeleton_estimates(.data, .alpha = 0.01)
}
\arguments{
\item{.data}{Input metabolic data.}

\item{.alpha}{Significance level threshold applied to each test.}
}
\value{
DAG skeleton object.
}
\description{
Uses the PC-algorithm and is mostly a wrapper around \code{\link[pcalg:skeleton]{pcalg::skeleton()}}.
}
\examples{

library(dplyr)
simulated_data \%>\%
select(contains("metabolite")) \%>\%
pc_skeleton_estimates()

}
