% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-ggraph.R, R/reexport-tibble.R,
%   R/reexport-tidyselect.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{scale_edge_width}
\alias{scale_edge_colour_gradient}
\alias{guide_edge_colourbar}
\alias{data_frame}
\alias{as_data_frame}
\alias{tibble}
\alias{as_tibble}
\alias{contains}
\alias{select_helpers}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{any_of}
\alias{starts_with}
\alias{last_col}
\alias{all_of}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggraph}{\code{\link[ggraph]{guide_edge_colourbar}}, \code{\link[ggraph:scale_edge_colour]{scale_edge_colour_gradient}}, \code{\link[ggraph]{scale_edge_width}}}

  \item{tibble}{\code{\link[tibble:deprecated]{as_data_frame}}, \code{\link[tibble]{as_tibble}}, \code{\link[tibble:deprecated]{data_frame}}, \code{\link[tibble]{tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect:all_of]{any_of}}, \code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{starts_with}}}
}}

