% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-netcoupler-functions.R
\name{rename_met}
\alias{rename_met}
\title{Rename feature names in order to avoid clash with glm.}
\usage{
rename_met(dat)
}
\arguments{
\item{dat}{Dataset of metabolites (columns) by samples (rows)}
}
\value{
Outputs a list of renamed column variables.
}
\description{
Rename feature names in order to avoid clash with glm.
}
