% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCoupler-package.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated dataset with an underlying Directed Graph structure for the metabolites.}
\format{
The simulated dataset is a \link[tibble:tibble-package]{tibble} with the
following variables:
\itemize{
\item Two outcome variables (\code{outcome_continuous} and \code{outcome_binary}) along
with survival time (\code{outcome_event_time}) that is used for the
\code{outcome_binary} variable
\item A generic \code{exposure} variable as continuous
\item 12 \verb{metabolite_*} variables
\item An \code{age} variable used as a confounder
}
}
\usage{
simulated_data
}
\description{
Simulated dataset with an underlying Directed Graph structure for the metabolites.
}
\keyword{datasets}
