% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{as_edge_tbl}
\alias{as_edge_tbl}
\title{Convert network graphs to edge tables as tibbles/data.frames.}
\usage{
as_edge_tbl(network_object)
}
\arguments{
\item{network_object}{Network graph from \code{\link[=nc_estimate_network]{nc_estimate_network()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble}, with at least two columns:
\itemize{
\item \code{source_node}: The starting node (variable).
\item \code{target_node}: The ending node (variable) that links with the source node.
\item \code{adjacency_weight}: (Optional) The "weight" given to the edge, which
represents the strength of the link between two nodes.
}
}
\description{
\lifecycle{experimental}
}
\seealso{
See \link{nc_estimate_links} for examples on using NetCoupler.
}
