% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_effects}
\alias{classify_effects}
\title{Classify direct, ambiguous, or no effect between exposure and network nodes.}
\usage{
classify_effects(.tbl)
}
\arguments{
\item{.tbl}{Multi-model estimates generated from \code{\link[=nc_exposure_estimates]{nc_exposure_estimates()}} or
\code{\link[=nc_outcome_estimates]{nc_outcome_estimates()}} that contain the model summaries.}
}
\value{
Outputs a \link[tibble:tibble-package]{tibble} with model estimates
between the exposure and the individual index network nodes, along with the
classification of direct effect or not.
}
\description{
\lifecycle{experimental}

Using the output of the \code{\link[=nc_exposure_estimates]{nc_exposure_estimates()}} or \code{\link[=nc_outcome_estimates]{nc_outcome_estimates()}},
classify whether the exposure variable has a direct, ambiguous, or no link
with the index node metabolic variable, conditional on neighbouring metablic
variables and on potential confounders.
}
\seealso{
See the \code{vignette("description")} for a
detailed description of the algorithm used to classify direct effects.
See \link{nc_model_estimates} for examples on using NetCoupler.
}
\keyword{internal}
