% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{nc_adjacency_graph}
\alias{nc_adjacency_graph}
\title{Compute the adjacency matrix of the graph with the data.}
\usage{
nc_adjacency_graph(.data, .graph)
}
\arguments{
\item{.data}{The data containing only the metabolic variables.}

\item{.graph}{The network graph object of the metabolic variable network.}
}
\value{
Outputs an \code{igraph} object from \code{\link[igraph:graph.adjacency]{igraph::graph.adjacency()}}.
}
\description{
\lifecycle{experimental}
}
\examples{

library(dplyr)
metabolite_data <- simulated_data \%>\%
  select(starts_with("metabolite"))
network <- metabolite_data \%>\%
  nc_create_network()
nc_adjacency_graph(metabolite_data, network) \%>\% class()
}
