% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{nc_estimate_network}
\alias{nc_estimate_network}
\title{Create an estimate of the metabolic network skeleton.}
\usage{
nc_estimate_network(data, cols = everything(), alpha = 0.05)
}
\arguments{
\item{data}{Data of the metabolic variables.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Variables to include
by using \code{\link[dplyr:select]{dplyr::select()}} style selection.}

\item{alpha}{The alpha level to set. Default is 0.05.}
}
\value{
Outputs a DAG skeleton.
}
\description{
\lifecycle{experimental}

Main NetCoupler network creator.
Estimates the skeleton based on a family of DAGs without specifying the direction of edges.
Defaults to using the PC algorithm to calculate possible edges.
}
\seealso{
See \link{nc_model_estimates} for examples on using NetCoupler.
}
