% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{nc_plot_network}
\alias{nc_plot_network}
\title{Title}
\usage{
nc_plot_network(.data, .graph)
}
\value{

}
\description{
Title
}
\examples{

library(dplyr)
library(NetCoupler)
metabolite_data <- simulated_data \%>\%
  select(starts_with("metabolite"))
network <- simulated_data \%>\%
  select(contains("metabolite")) \%>\%
  nc_create_network()
nc_plot_network(metabolite_data, network)

}
