% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{nc_plot_network}
\alias{nc_plot_network}
\title{Plot of the network of the metabolic variables.}
\usage{
nc_plot_network(
  data,
  edge_tbl,
  edge_label_threshold = 0.2,
  fn_node_rename = NULL
)
}
\arguments{
\item{data}{The data that was also used to generate the \code{\link[=nc_estimate_network]{nc_estimate_network()}}.}

\item{edge_tbl}{The network graph object of the metabolic variable network.}

\item{edge_label_threshold}{Threshold set for edge weight value above which
the edge label will be kept. This argument helps to reduce clutter in the
graph.}

\item{fn_node_rename}{Function to pass to rename the metabolic variables.
Preferably use functions that search and replace patterns, like \code{\link[=gsub]{gsub()}} or
\code{\link[stringr:str_replace]{stringr::str_replace()}}.}
}
\value{
Outputs a \code{ggplot2} object of the metabolic network.
}
\description{
\lifecycle{experimental}
}
\seealso{
See \link{nc_estimate_links} for examples on using NetCoupler.
}
