% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCoupler-package.R
\docType{package}
\name{NetCoupler-package}
\alias{NetCoupler}
\alias{NetCoupler-package}
\title{NetCoupler: Inference of Causal Links Between Metabolomics and Disease Incidence}
\description{
The NetCoupler algorithm identifies direct effects of correlated, 
    high-dimensional exposures on time-to-event data. It can also be used with
    an external exposure that may effect the -omic network (e.g. lifestyle), to
    be used to identify indirect pathways between the external exposure and an 
    outcome.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ClemensWittenbecher/NetCoupler}
  \item Report bugs at \url{https://github.com/ClemensWittenbecher/NetCoupler/issues}
}

}
\author{
\strong{Maintainer}: Clemens Wittenbecher \email{Clemens.Wittenbecher@dife.de}

Authors:
\itemize{
  \item Luke Johnston \email{lwjohnst@gmail.com}
}

Other contributors:
\itemize{
  \item Helena Zacharias \email{helena.zacharias@helmholtz-muenchen.de} [contributor]
}

}
\keyword{internal}
