% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tibble.R, R/reexport-tidyselect.R,
%   R/reexports-ggraph.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{data_frame}
\alias{as_data_frame}
\alias{tibble}
\alias{as_tibble}
\alias{contains}
\alias{select_helpers}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{last_col}
\alias{vars}
\alias{scale_edge_width}
\alias{scale_edge_colour_gradient}
\alias{guide_edge_colourbar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{vars}}}

  \item{ggraph}{\code{\link[ggraph]{guide_edge_colourbar}}, \code{\link[ggraph]{scale_edge_colour_gradient}}, \code{\link[ggraph]{scale_edge_width}}}

  \item{tibble}{\code{\link[tibble]{as_data_frame}}, \code{\link[tibble]{as_tibble}}, \code{\link[tibble]{data_frame}}, \code{\link[tibble]{tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{last_col}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}}
}}

