% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCoupler-package.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated dataset with an underlying Directed Graph structure for the metabolites.}
\format{The simulated dataset, output as a \link[tibble:tibble-package]{tibble}
contains two outcome variables (\code{survival_time} and \code{case_status}), one
generic \code{exposure}, 12 \verb{metabolite_*} variables, and an \code{age} variable.}
\usage{
simulated_data
}
\description{
Simulated dataset with an underlying Directed Graph structure for the metabolites.
}
\keyword{datasets}
